+++ packages/askar/src/index.ts
@@ -4,6 +4,7 @@ export {
   AskarWalletPostgresStorageConfig,
   AskarWalletPostgresConfig,
   AskarWalletPostgresCredentials,
+  AskarProfileWallet,
 } from './wallet'
 
 // Storage
@@ -11,3 +12,4 @@ export { AskarStorageService } from './storage'
 
 // Module
 export { AskarModule } from './AskarModule'
+export { AskarModuleConfigOptions, AskarMultiWalletDatabaseScheme } from './AskarModuleConfig'
